KokkosBatched - gemv host
